requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow-cpu==2.15.0
pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
gunicorn==21.2.0
requests==2.31.0
tqdm==4.66.1
